PARQUET-400: Remove Hadoop dependencies from SeekableInputStream.
