+++ parquet-hadoop/src/main/java/org/apache/parquet/hadoop/util/SeekableInputStream.java
@@ -19,48 +19,11 @@
 
 package org.apache.parquet.hadoop.util;
 
-import org.apache.hadoop.fs.FSDataInputStream;
-import org.apache.parquet.Log;
-import org.apache.parquet.io.ParquetDecodingException;
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
 import java.nio.ByteBuffer;
 
 public abstract class SeekableInputStream extends InputStream {
-
-  private static final Log LOG = Log.getLog(SeekableInputStream.class);
-
-  private static final Class<?> byteBufferReadableClass;
-
-  private static final Constructor<SeekableInputStream> h2SeekableConstructor;
-
-  static {
-    byteBufferReadableClass = getReadableClass();
-    h2SeekableConstructor = getH2SeekableConstructor();
-  }
-
-  public static SeekableInputStream wrap(FSDataInputStream stream) {
-    if (byteBufferReadableClass != null && h2SeekableConstructor != null &&
-        byteBufferReadableClass.isInstance(stream) &&
-        supportsByteBufferReads(stream)) {
-      try {
-        return h2SeekableConstructor.newInstance(stream);
-      } catch (InstantiationException e) {
-        LOG.warn("Could not instantiate H2SeekableInputStream, falling back to byte array reads", e);
-        return new H1SeekableInputStream(stream);
-      } catch (IllegalAccessException e) {
-        LOG.warn("Could not instantiate H2SeekableInputStream, falling back to byte array reads", e);
-        return new H1SeekableInputStream(stream);
-      } catch (InvocationTargetException e) {
-        throw new ParquetDecodingException(
-            "Could not instantiate H2SeekableInputStream", e.getTargetException());
-      }
-    } else {
-      return new H1SeekableInputStream(stream);
-    }
-  }
 
   public abstract long getPos() throws IOException;
 
@@ -74,54 +37,4 @@ public abstract class SeekableInputStream extends InputStream {
 
   public abstract void readFully(ByteBuffer buf) throws IOException;
 
-  private static final ByteBuffer ZERO_LEN_BYTE_BUFFER = ByteBuffer.wrap(new byte[0]);
-
-  private static boolean supportsByteBufferReads(FSDataInputStream stream) {
-    // FSDataInputStream implements ByteBufferReadable, but may throw
-    // UnsupportedOperationException if it isn't actually supported. This tests
-    // whether the method throws by trying to read a 0-length buffer, which has
-    // no effect on the stream when it is supported.
-    try {
-      stream.read(ZERO_LEN_BYTE_BUFFER);
-      return true;
-    } catch (UnsupportedOperationException e) {
-      return false;
-    } catch (IOException e) {
-      return false;
-    }
-  }
-
-  private static Class<?> getReadableClass() {
-    try {
-      return Class.forName("org.apache.hadoop.fs.ByteBufferReadable");
-    } catch (ClassNotFoundException e) {
-      return null;
-    } catch (NoClassDefFoundError e) {
-      return null;
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Class<SeekableInputStream> getH2SeekableClass() {
-    try {
-      return (Class<SeekableInputStream>) Class.forName(
-          "org.apache.parquet.hadoop.util.H2SeekableInputStream");
-    } catch (ClassNotFoundException e) {
-      return null;
-    } catch (NoClassDefFoundError e) {
-      return null;
-    }
-  }
-
-  private static Constructor<SeekableInputStream> getH2SeekableConstructor() {
-    Class<SeekableInputStream> h2SeekableClass = getH2SeekableClass();
-    if (h2SeekableClass != null) {
-      try {
-        return h2SeekableClass.getConstructor(FSDataInputStream.class);
-      } catch (NoSuchMethodException e) {
-        return null;
-      }
-    }
-    return null;
-  }
 }
